RF: Introduced textures and scene
- Created a file for texture related things in v2.io
- Created a file for scene extraction in v2.window
- Moved the old data in scene.py to renderer.py
